+ poc model split grammar vs semantics
