+++ My1kWordsEe/Models/Grammar/Number.cs
@@ -1,0 +1,12 @@
+using System.Text.Json.Serialization;
+
+namespace My1kWordsEe.Models.Grammar
+{
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public enum Number
+    {
+        None, // Default value | cannot be determined
+        Singular,
+        Plural
+    }
+}
