+++ My1kWordsEe/Models/Semantics/SampleSentence.cs
@@ -1,0 +1,18 @@
+
+using System.Text.Json.Serialization;
+
+namespace My1kWordsEe.Models.Grammar
+{
+    /// <summary>
+    /// A sample sentence illustrating the use of a given Estonian word form.
+    /// </summary>
+    public record SampleSentence
+    {
+        // word?
+
+        public required IDictionary<LanguageCode, string> Sentence { get; init; } = new Dictionary<LanguageCode, string>();
+        public required string EnSentence { get; init; }
+        public Uri? EeAudioUrl { get; init; }
+        public Uri? ImageUrl { get; init; }
+    }
+}
